Update time_calculator.py
Woooo! Finally complete! Add your start time for in for the first argument and the length of time (hr:min) to the second argument and it'll tell you what time it'll be in 12hr format. There is a third, optional argument for if you, also, want to calculate the weekday, but you have to enter the day of the week you'd like to calculate from ( add_time("1:45 PM", "1879:25", "Monday").